requirements.txt
gym==0.17.3
matplotlib==3.3.3
atari-py==0.2.6
stable-baselines3[extra]
gym_anytrading
QuantStats==0.0.34
optuna
python-dotenv==0.18.0
plotly==5.3.1
kaleido

